Move pull_model_trajectories to separate file
