+++ examples/art-e/art_e/evaluate/llm_judge_charts.py
@@ -1,9 +1,7 @@
 # %%
 
-from art.utils.benchmarking.load_trajectories import (
-    load_trajectories,
-    pull_model_trajectories,
-)
+from art.utils.benchmarking.load_trajectories import load_trajectories
+from art.utils.benchmarking.pull_model_trajectories import pull_model_trajectories
 from art_e.evaluate.charts import training_progress_chart, comparison_models_bar_chart
 from all_experiments import models
 import asyncio
